requirements.txt
sentence-transformers
pinecone
python-docx
PyPDF2
nltk
streamlit
pandas